requirements.txt
asgiref==3.9.1
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.8.0
Django==5.2.4
django-environ==0.12.0
fonttools==4.58.5
pillow==11.3.0
psycopg==3.2.9
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1